Improve logging and add confirmation before deleting raffle

+++ app/commands.py
@@ -12,22 +12,19 @@ def delete(raffle_id):
     try:
         raffle = Raffle.query.filter_by(submission_id=raffle_id).one()
     except:
-        click.echo('Raffle with ID \'{}\' does not exist.'.format(raffle_id))
+        click.echo('Raffle \'{}\' does not exist.'.format(raffle_id))
         return
 
-    current_app.logger.info('Removing {}'.format(raffle_id))
-    click.echo('Removing {}'.format(raffle_id))
-
-    try:
-        db.session.delete(raffle)
-        db.session.commit()
-        current_app.logger.exception(
-            'Successfully removed {}'.format(raffle_id)
-        )
-        click.echo('Successfully removed {}'.format(raffle_id))
-    except:
-        db.session.rollback()
-        current_app.logger.exception(
-            'Error while trying to remove {}'.format(raffle_id)
-        )
-        click.echo('Something went wrong while trying to delete the raffle.')
+    if click.confirm('Really delete Raffle {}?'.format(raffle_id)):
+        current_app.logger.info('[COMMAND] Removing {}'.format(raffle_id))
+        try:
+            db.session.delete(raffle)
+            db.session.commit()
+            current_app.logger.info('[COMMAND] Removed {}'.format(raffle_id))
+            click.echo('Successfully removed {}'.format(raffle_id))
+        except:
+            db.session.rollback()
+            current_app.logger.exception(
+                '[COMMAND] Error while trying to remove {}'.format(raffle_id)
+            )
+            click.echo('Something went wrong while deleting the raffle.')
